refactor/get_rating->rating
