+++ src/app/interfaces/routes.rs
@@ -3,7 +3,7 @@
 use tonic::transport::server::Router;
 use tonic_reflection::server::{ServerReflection, ServerReflectionServer};
 
-use crate::features::{chart, get_rating, user};
+use crate::features::{chart, rating, user};
 
 /// Creates a new default reflection server for this app
 pub fn build_reflection_service() -> ServerReflectionServer<impl ServerReflection> {
@@ -19,7 +19,7 @@ pub fn build_reflection_service() -> ServerReflectionServer<impl ServerReflectio
 /// the [`Router`] won't be otherwise modified.
 pub fn build_servers<R>(router: Router<R>) -> Router<R> {
     let user_service = user::service::build_service();
-    let app_service = get_rating::service::build_service();
+    let app_service = rating::service::build_service();
     let chart_service = chart::service::build_service();
 
     router
