+++ src/features/mod.rs
@@ -1,7 +1,7 @@
 //! Contains various feature implementations for the ratings backend.
 
 pub mod chart;
-pub mod get_rating;
+pub mod rating;
 pub mod user;
 
 mod common;
