+++ src/features/rating/errors.rs
