+++ src/features/rating/service.rs
